auth/src/requirements/prod.txt
redis==5.2.0
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pydantic-settings~=2.6.0
aiohttp~=3.10.0
orjson~=3.10.0
fastapi-cache2~=0.2.2
uvicorn>=0.32.0,<0.33.0
sqlalchemy~=2.0.36
psycopg~=3.2.3
psycopg-binary~=3.2.3
werkzeug~=3.1.3
greenlet~=3.1.1
asyncpg~=0.30.0
alembic>=1.14.0
PyJWT==2.10.1
slowapi>=0.1.9
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0